requirements.txt
python-telegram-bot==20.3
ryanair-py==3.0.0
python-dotenv==1.0.0
requests